feat(nextjs): Update protect() to redirect if handling a page request
We're leveraging the isPageRequest and isServerActionRequest checks to detect whether the current request is a page request. For page requests, instead of throwing a 404 or 401, we're redirecting to SIGN_IN_URL automatically.

In order to achieve that, we're using native headers but also the `next-urk` and `next-action` headers of NextJS